Admin Statistics System Bugs Fixed && GUI Done

+++ Users/Admins/AdminStatisticsSystem/Receivers/ReceiverStrategyPattern/Products/ProductAveragePriceDefault.cs
@@ -18,7 +18,7 @@ namespace OnlineStore.Users.Admins.AdminsStatisticsSystemCommands.Receivers.Rece
 
         public String Average()
         {
-            double res1 = Convert.ToInt32(query.StorePricesSum());
+            double res1 = Convert.ToDouble(query.StorePricesSum());
             int res2 = Convert.ToInt32(query.StorePricesCnt());
             return "Store prices Average: " + ((double)res1 / res2).ToString();
         }
